+++ packages/std-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mud-classic/std-client
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @mud-classic/network@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes